Use cmake 2 (#1)
* add project to cmake build

* remove meson files

* link chlib with m

* modify to compile on msvc

* make modifications for linux build

* fix memory testing

* remove meson tasks and settings

* skip mem tests for emscripten

* update tests for ASAN/LSAN cases

+++ src/memory.c
@@ -1,13 +1,13 @@
 #include <stdlib.h>
 
 void *
-chl_alloc (long nbytes)
+chl_alloc (size_t nbytes)
 {
   return malloc (nbytes);
 }
 
 void *
-chl_calloc (long count, long nbytes)
+chl_calloc (size_t count, size_t nbytes)
 {
   return calloc (count, nbytes);
 }
